deep-fake-detector/requirements.txt
livekit-agents[openai,silero,deepgram,rime]~=1.0
python-dotenv~=1.0 